chore: Chart context menu permissions cleanup (#30854)
